% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_letter.R
\name{cover_letter}
\alias{cover_letter}
\alias{templ_cover_letter}
\title{Resume Templates}
\usage{
cover_letter(...)

templ_cover_letter()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for Cover Letter documents
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab title of the document, appears on first row \cr
   \code{author} \tab name of the author, appears on second row \cr
   \code{affiliation} \tab institutional affiliation, appears on third row \cr
   \code{email} \tab your email, for the fourth row \cr
}
}

