% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curriculum_vitae.R
\name{curriculum_vitae}
\alias{curriculum_vitae}
\alias{templ_curriculum_vitae}
\title{Resume Templates}
\usage{
curriculum_vitae(...)

templ_curriculum_vitae()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for Curriculum Vitae documents
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab title of the document, appears on first row \cr
   \code{author} \tab name of the author, appears on second row \cr
   \code{affiliation} \tab institutional affiliation, appears on third row \cr
   \code{email} \tab your email, for the fourth row \cr
}
}

\examples{
path = "Untitled"
if (file.exists(path)) {unlink(path, recursive = TRUE)}
rmarkdown::draft(paste0(path, ".Rmd"),
template="curriculum_vitae", 
package="resume", 
create_dir = TRUE, 
edit = FALSE)
rmarkdown::render(paste0(path, "/", path, ".Rmd"))
if (file.exists(path)) {unlink(path, recursive = TRUE)}
}
