% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motivatiebrief.R, R/resume.R
\name{templ_motivatiebrief}
\alias{templ_motivatiebrief}
\alias{resume}
\alias{templ_resume}
\title{Resume}
\usage{
templ_motivatiebrief()

resume(...)

templ_resume()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for Résumé documents
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{name} \tab your name and surname \cr
   \code{contact} \tab address, telephone,  email, web \cr
   \code{profile} \tab short profile summary \cr
   \code{...} \tab the rest \cr
}
}

\examples{
path = "Untitled"
if (file.exists(path)) {unlink(path, recursive = TRUE)}
rmarkdown::draft(paste0(path, ".Rmd"),
template="resume", 
package="resume", 
create_dir = TRUE, 
edit = FALSE)
if(Sys.info()["sysname"]  != "Windows") {
str = readLines(paste0(path, "/", path, ".Rmd"))
str = gsub("Arial", "Liberation Serif", str )
writeLines(str, paste0(path, "/", path, ".Rmd")) }
rmarkdown::render(paste0(path, "/", path, ".Rmd"))
if (file.exists(path)) {unlink(path, recursive = TRUE)}

}
