% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_ats.R
\name{resume_ats}
\alias{resume_ats}
\title{Resume ATS}
\usage{
resume_ats(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for ATS friendly Résumé documents
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{name} \tab your name and surname \cr
   \code{contact} \tab address, telephone,  email, web \cr
   \code{profile} \tab short profile summary \cr
   \code{...} \tab the rest \cr
}
}

