% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav.R
\name{resume_navbar}
\alias{resume_navbar}
\title{A Resume Navbar}
\usage{
resume_navbar(refs = c(a = "A"), image = "img/profile.jpg",
  color = "#bd5d38")
}
\arguments{
\item{refs}{a named list with names being the ID and Elements being the text displayed.}

\item{image}{Image to put in the nav.}

\item{color}{Color of the sidebar}
}
\value{
An HTML taglist
}
\description{
A Resume Navbar
}
