% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{resume_page}
\alias{resume_page}
\title{Resume page}
\usage{
resume_page(app_title, nav, body)
}
\arguments{
\item{app_title}{Title of the App}

\item{nav}{A navbar, constructed with \code{resume_navbar}}

\item{body}{A body, constructed with \code{resume_body} & \code{resume_sections}}
}
\value{
An HTML taglist
}
\description{
A Shinny App with the Bootstrap Resume Template
}
