% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{resume_section}
\alias{resume_section}
\title{A section in the page}
\usage{
resume_section(id, ...)
}
\arguments{
\item{id}{ID of the sections, should reflect the \code{resume_navbar} structure}

\item{...}{elements to put inside the section}
}
\value{
An HTML taglist
}
\description{
A section in the page
}
