% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ua_list_motywacyjny.R
\name{ua_list_motywacyjny}
\alias{ua_list_motywacyjny}
\alias{templ_ua_list_motywacyjny}
\title{ua_list_motywacyjny}
\usage{
ua_list_motywacyjny(...)

templ_ua_list_motywacyjny()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for list_motywacyjny documents
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab title of the document, appears on first row \cr
   \code{author} \tab name of the author, appears on second row \cr
   \code{affiliation} \tab institutional affiliation, appears on third row \cr
   \code{email} \tab your email, for the fourth row \cr
}
}

\examples{
path = "Untitled"
if (file.exists(path)) {unlink(path, recursive = TRUE)}
rmarkdown::draft(paste0(path, ".Rmd"),
template="ua_list_motywacyjny", 
package="resume", 
create_dir = TRUE, 
edit = FALSE)
rmarkdown::render(paste0(path, "/", path, ".Rmd"))
if (file.exists(path)) {unlink(path, recursive = TRUE)}
}
