% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_letter.R, R/list_motywacyjny.R
\name{cover_letter}
\alias{cover_letter}
\alias{templ_cover_letter}
\alias{templ_list_motywacyjny}
\title{Cover Letter}
\usage{
cover_letter(...)

templ_cover_letter()

templ_list_motywacyjny()
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
A template for Cover Letter documents
}
\section{About YAML header fields}{
This section documents some of the YAML fields to know
for this template.\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab title of the document, appears on first row \cr
   \code{author} \tab name of the author, appears on second row \cr
   \code{affiliation} \tab institutional affiliation, appears on third row \cr
   \code{email} \tab your email, for the fourth row \cr
}
}

\examples{
path = "Untitled"
if (file.exists(path)) {unlink(path, recursive = TRUE)}
rmarkdown::draft(paste0(path, ".Rmd"),
template="cover_letter", 
package="resume", 
create_dir = TRUE, 
edit = FALSE)
rmarkdown::render(paste0(path, "/", path, ".Rmd"))
if (file.exists(path)) {unlink(path, recursive = TRUE)}
}
