% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_letter.R
\name{cover_letter}
\alias{cover_letter}
\title{Template to build a Cover letter.}
\usage{
cover_letter(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
First create a new R Markdown document from template Cover Letter then modify it 
and knit PDF output.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "cover_letter", package = "resumeR")
}

}
