% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curriculum_vitae.R
\name{curriculum_vitae}
\alias{curriculum_vitae}
\title{Template to build Curriculum Vitae.}
\usage{
curriculum_vitae(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
First create a new R Markdown document from template Curriculum Vitae then modify it 
and knit PDF output.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyCV.Rmd", template = "curriculum_vitae", package = "resume")
}

}
