% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{resume}
\alias{resume}
\title{Template to build Resume.}
\usage{
resume(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
First create a new R Markdown document from template Resume then modify it 
and knit PDF output.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "resume", package = "resume")
}

}

